Delete initclaim window.BASE_URL in favor of local .env file (#50)
Background: the original `BASE_URL` was a workaround for local development since django and react dev servers are listening on different ports and need to redirect to each other. That only affects dev, since otherwise the React app is built and served via Django, meaning they share a base hostname/url.

Since we added the "proxy" feature for API XHR, the `BASE_URL` was only getting used for the initial redirect and we can use the `.env` file for that instead of the more intrusive/complex template-level `window` variable.